+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Changelog for the gruf-balancer gem.
 
 ### Pending release
 
+- Allow configuring the client class when adding a client
+
 ### 0.0.1
 
 - Initial release